% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise_content.R
\name{render_lectures}
\alias{render_lectures}
\title{Deploy the lecture slides in HTML format}
\usage{
render_lectures(lecture)
}
\arguments{
\item{lecture}{character object indicating which lecture to visualise:
\itemize{
  \item{'Lecture1': }{Functions and parallel computation}
  \item{'Lecture2': }{Advanced spatial and temporal statistics}
  \item{'Lecture3': }{Machine learning algorithms in R - I}
  \item{'Lecture4': }{Machine learning algorithms in R - II}
  \item{'Lecture5': }{Working with R Markdown and Leaflet}
}}
}
\value{
this function directs the user to the selected lecture.
}
\description{
Deploy the lecture slides in HTML format
}
